requirements.txt
requests>=2.28.0
PyYAML>=6.0
feedparser>=6.0.0
Jinja2>=3.1.0
openai>=1.0.0
anthropic>=0.18.0
